fix: align HAL with audited WS63 register contracts

+++ tests/hil/uart.rs
@@ -1,7 +1,7 @@
 use crate::{hal, pac};
 
 #[cfg(feature = "chip-ws63")]
-fn uart0_dividers() -> (u16, u16, u16) {
+fn uart0_dividers() -> (u32, u32, u32) {
     // DIV_L/DIV_H share the UART data-register addresses and are only visible
     // while divisor-latch access is enabled. Preserve the configured line
     // control value around the diagnostic readback.
@@ -24,9 +24,9 @@ pub(crate) fn uart0_divider_config() {
     let pclk = hal::soc::chip::UART_CLOCK_HZ;
     let div64 = ((pclk as u64) * 4 / (cfg.baudrate.baud() as u64)) as u32;
     let div = div64 >> 6;
-    let exp_div_fra = (div64 & 0x3F) as u16;
-    let exp_div_l = (div & 0xFF) as u16;
-    let exp_div_h = ((div >> 8) & 0xFF) as u16;
+    let exp_div_fra = div64 & 0x3F;
+    let exp_div_l = div & 0xFF;
+    let exp_div_h = (div >> 8) & 0xFF;
 
     let (div_l, div_h, div_fra) = uart0_dividers();
     assert_eq!(div_l, exp_div_l, "UART0 div_l mismatch");
@@ -50,9 +50,9 @@ pub(crate) fn uart0_boot_clock_divider_config() {
     );
     let div64 = ((pclk as u64) * 4 / (cfg.baudrate.baud() as u64)) as u32;
     let div = div64 >> 6;
-    let exp_div_fra = (div64 & 0x3F) as u16;
-    let exp_div_l = (div & 0xFF) as u16;
-    let exp_div_h = ((div >> 8) & 0xFF) as u16;
+    let exp_div_fra = div64 & 0x3F;
+    let exp_div_l = div & 0xFF;
+    let exp_div_h = (div >> 8) & 0xFF;
 
     let (div_l, div_h, div_fra) = uart0_dividers();
     assert_eq!(div_l, exp_div_l, "UART0 boot div_l mismatch");
